refactor: reorganize project structure and rename directories - Moved files to their appropriate directories - Renamed directories
