+++ questao_06/Questao6.java
